feat: Add Arithmetics class with unnest_expression method
The code changes introduce a new `Arithmetics` class that includes the `unnest_expression` method. This method takes an expression as input and returns a list of operations. The operations are represented as tuples, with each tuple containing the variable name, function name, and arguments.

Recent user commits:
- Verbalization templates added
- Renaming some funcs to capture some aspect of their function
- unnest ready
- copy geo verb

Recent repository commits:
- Merge branch 'main' of github.com:felixludos/alphageometry
- Verbalization templates added
- Renaming some funcs to capture some aspect of their function
- unnest ready
- copy geo verb
- Merge pull request #1 from felixludos/nl_verbalization
- Merge branch 'nl_verbalization'
- Renaming
- Now operations are classes. Interventional gen is functioning
- Creating a new formal system based on Arithmetic